update Test_SocketRequest
InputStream → InputStreamReader → BufferedReader .
Why：
InputStream.available() ：  Note that while some implementations of
InputStream will return the total number of bytes in the stream, many
will not. It is never correct to use the return value of this method to
allocate a buffer intended to hold all data in this stream.